legal-rag-system/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
python-dotenv==1.0.0
ollama>=0.1.0
supabase
openai>=1.10